Fix BPF code and get rid of .BTF stripping
The work-around I had was stale, not loading the .BTF section resulted
in not having the verifier find the bug I had in accessing a map member
with BPF_CORE_READ. However that worked (not having BTF), wiht that
fixed and the harmful strip gone global variables are possible and also
the RewriteConstants works fine (alternative way to pass values from
golang -> BPF). Hurray.

The only thing left is to find out if the global is usable from
preprocessor (probably not).